DataExtractorWorkerRole/requirements.txt
bs4
azure-storage